pyproject.toml
[tool.poetry]
name = "weedly-bot"
version = "1"
description = "telegram bot for weedly"
authors = ["kbondar17 <ironbondar@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "^2.16"
python-telegram-bot-pagination = "^0.0.2"
httpx = "^0.21.3"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
